backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-docx
python-pptx
google-generativeai
python-dotenv
python-multipart
